+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+ * `FreezeType`
+ * `FreezeTransaction.[get|set]FreezeType()`
+ * `FreezeTransaction.[get|set]FileId()`
+ * `FreezeTransaction.[get|set]FileHash()`
+
+### Deprecated
+
+ * `FreezeTransaction.[get|set]UpdateFileId()`, use `.[get|set]FileId()` instead.
+ * `FreezeTransaction.[get|set]UpdateFileHash()`, use `.[get|set]FileHash()` instead.
+
 ## v2.2.0-beta.2
 
 ### Fixed